+++ Practice_009/Ad/UI/PatientDetailTabListViewController/PatientDetailTabListProtocol.swift
@@ -8,7 +8,7 @@ protocol PatientDetailTabListViewProtocol: BaseViewControllerProtocol {
 }
 
 protocol PatientDetailTabListPresenterProtocol: NSObjectProtocol {
-    func postPatientCoach(userId: String, timeStamp: String, borgUUID: String, patientCoachList: [(speed: Int?, time: Int?)])
+    func postPatientCoach(userId: String, timeStamp: String, borgUUID: String, patientCoachList: [(speed: Double?, time: Double?)])
     func getPatientSurvey(userId: String, timestamp: String)
     func getPatientCoach(userId: String, timestamp: String, borgUUID: String)
     func getPatientBorg(userId: String, timestamp: String)
